+++ Software/Testing/micropython/software/main.py
@@ -12,311 +12,22 @@
 #   "mic_left_channel_16bits.wav"
 #
 # Hardware tested:
-# - INMP441 microphone module 
+# - INMP441 microphone module
 # - MSM261S4030H0 microphone module
 
 import os
-from machine import Pin
-from sdcard import SDCard
-from machine import I2S
-from machine import SPI
 import time
 from ili9341 import color565
-from xglcd_font import XglcdFont
-from setup import sd_spi, display, sd, rotary_enc, neopixels, button
-
-
-# Load font
-unispace = XglcdFont('fonts/Unispace12x24.c', 12, 24)
-
-# Function to scroll through a list of menu items and return a selection on encoder press
-def menu_select(last_position, menu_items):
-    # Force last_position to not equal rotary_enc.value() and be % = 0
-    last_position = -len(menu_items)
-    rotary_enc.reset()
-    item_selected = False
-    while item_selected is False:
-        current_position = rotary_enc.value()
-
-        # Generate a valid index from the position
-        if current_position != last_position:
-            index = current_position % len(menu_items)
-            # Display item
-            pretty_name = menu_items[index]["pretty"]
-            text = str.format("{}: {}", index, pretty_name)
-            display.draw_text(0, 108, text, unispace,
-                  color565(0, 0, 0), background = color565(255,255,255))
-            last_position = current_position
-
-        # Select item
-        #enc_buttons_event = enc_buttons.events.get()
-        #if enc_buttons_event and enc_buttons_event.pressed:
-        #    index = current_position % len(menu_items)
-        #    return menu_items[index]["name"]
-
-
-def show_menu(menu, highlight, shift):
-    """Shows the menu on the screen"""
-
-    # menu variables
-    item = 1
-    line = 1
-    line_height = 24
-    offset = 5
-    total_lines = 5
-
-    # Shift the list of files so that it shows on the display
-    short_list = []
-    for index in range(shift, shift + total_lines):
-        try:
-            short_list.append(menu[index]["pretty"])
-        except IndexError:
-            print("show_menu: Bad Index")
-    for item in short_list:
-        if highlight == line:
-            display.draw_text(0, (line-1) * line_height, item, unispace,
-                  color565(0, 0, 0), background = color565(255,255,255))
-
-        else:
-            display.draw_text(0, (line-1) * line_height, item, unispace,
-                  color565(255, 255, 255), background = color565(0, 0, 0))
-        line += 1
-
-
-class State(object):
-    def __init__(self):
-        pass
-
-    @property
-    def name(self):
-        return ""
-
-    def enter(self, core_machine):
-        pass
-
-    def exit(self, core_machine):
-        pass
-
-    def update(self, core_machine):
-        return True
-
-
-class StateMachine(object):
-    def __init__(self):
-        self.state = None
-        self.states = {}
-        self.last_enc1_pos = rotary_enc.value()
-        self.paused_state = None
-        self.ticks_ms = 0
-        self.animation = None
-
-    def add_state(self, state):
-        self.states[state.name] = state
-
-    def go_to_state(self, state_name):
-        if self.state:
-            self.state.exit(self)
-        self.state = self.states[state_name]
-        self.state.enter(self)
-
-    def update(self):
-        if self.state:
-            self.state.update(self)
-            if self.ticks_ms > 0:
-                time.sleep(self.ticks_ms / 1000)
-
-    # When pausing, don't exit the state
-    def pause(self):
-        self.state = self.states["paused"]
-        self.state.enter(self)
-
-    # When resuming, don't re-enter the state
-    def resume_state(self, state_name):
-        if self.state:
-            self.state.exit(self)
-        self.state = self.states[state_name]
-
-
-class StartupState(State):
-    color = (0, 0, 0)
-    timer = 0
-    stage = 0
-
-    @property
-    def name(self):
-        return "startup"
-
-    def enter(self, machine):
-        display.clear()
-        State.enter(self, machine)
-
-    def exit(self, machine):
-        State.exit(self, machine)
-
-    def update(self, machine):
-        # Display Test
-        display.draw_image("dczia.raw")
-        time.sleep(4)
-        display.draw_text(0, 0, 'DCZia SSTV Badge Test', unispace,
-                          color565(255, 255, 255))
-
-        # Neopixel Test
-        for value in range(0,18):
-            neopixels[value] = [25, 0, 0]
-        neopixels.write()
-        time.sleep(1)
-
-        for value in range(0,18):
-            neopixels[value] = [0, 25, 0]
-        neopixels.write()
-        time.sleep(1)
-
-        for value in range(0,18):
-            neopixels[value] = [0, 0, 25]
-        neopixels.write()
-        core_machine.go_to_state("menu")
-
-class MenuState(State):
-    menu_items = [
-        {
-            "name": "party",
-            "pretty": "Party",
-        },
-        {
-            "name": "sstv_encoder",
-            "pretty": "SSTV Encoder",
-        },
-        {
-            "name": "sstv_decoder",
-            "pretty": "SSTV Decoder",
-        },
-        {
-            "name": "image_display",
-            "pretty": "Display Image",
-        },
-        {
-            "name": "startup",
-            "pretty": "Startup State (test)",
-        },
-    ]
-
-    @property
-    def name(self):
-        return "menu"
-
-    def __init__(self):
-        self.total_lines = 10
-        self.list_length = len(self.menu_items)
-        self.highlight = 1
-        self.shift = 0
-
-    def enter(self, machine):
-        display.clear()
-        self.last_position = 0
-        self.button_position = 1
-        show_menu(self.menu_items, self.highlight, self.shift)
-        State.enter(self, machine)
-
-    def exit(self, machine):
-        rotary_enc.reset()
-        State.exit(self, machine)
-
-    def update(self, machine):
-        # Some code here to use an encoder to scroll through menu options, press to select one
-        position = rotary_enc.value()
-        list_length = len(self.menu_items)  # TODO: Doesn't change every frame
-
-        if self.last_position != position:
-            if position < self.last_position:
-                if self.highlight > 1:
-                    self.highlight -= 1
-                else:
-                    if self.shift > 0:
-                        self.shift -= 1
-            else:
-                if self.highlight < self.total_lines:
-                    self.highlight += 1
-                else:
-                    if self.shift + self.total_lines < list_length:
-                        self.shift += 1
-            show_menu(self.menu_items, self.highlight, self.shift)
-        self.last_position = position
-        if button.value() is 0:
-            machine.go_to_state(
-                self.menu_items[self.highlight - 1 + self.shift]["name"]
-            )
-
-class PartyState(State):
-    @property
-    def name(self):
-        return "party"
-
-    def enter(self, machine):
-        display.clear()
-        display.draw_text(0, 0, 'Party Mode', unispace,
-                          color565(255, 255, 255))
-        State.enter(self, machine)
-
-    def exit(self, machine):
-        State.exit(self, machine)
-
-    def update(self, machine):
-        if button.value() is 0:
-            machine.go_to_state("menu")
-
-class SSTVEncoderState(State):
-    @property
-    def name(self):
-        return "sstv_encoder"
-
-    def enter(self, machine):
-        display.clear()
-        display.draw_text(0, 0, 'SSTV Encoder', unispace,
-                          color565(255, 255, 255))
-        State.enter(self, machine)
-
-    def exit(self, machine):
-        State.exit(self, machine)
-
-    def update(self, machine):
-        if button.value() is 0:
-            machine.go_to_state("menu")
-
-class SSTVDecoderState(State):
-    @property
-    def name(self):
-        return "sstv_decoder"
-
-    def enter(self, machine):
-        display.clear()
-        display.draw_text(0, 0, 'SSTV Decoder', unispace,
-                          color565(255, 255, 255))
-        State.enter(self, machine)
-
-    def exit(self, machine):
-        State.exit(self, machine)
-
-    def update(self, machine):
-        if button.value() is 0:
-            machine.go_to_state("menu")
-
-class ImageDisplayState(State):
-    @property
-    def name(self):
-        return "image_display"
-
-    def enter(self, machine):
-        display.clear()
-        display.draw_image("dczia.raw")
-        State.enter(self, machine)
-
-    def exit(self, machine):
-        State.exit(self, machine)
-
-    def update(self, machine):
-        if button.value() is 0:
-            machine.go_to_state("menu")
-
+from machine import Pin, I2S, SPI
+from sdcard import SDCard
+from setup import sd_spi, display, sd, rotary_enc, neopixels, button, unispace
+from state import State, StateMachine
+from state_image import ImageDisplayState
+from state_menu import MenuState
+from state_party import PartyState
+from state_sstv_decode import SSTVDecoderState
+from state_sstv_encode import SSTVEncoderState
+from state_startup import StartupState
 
 core_machine = StateMachine()
 core_machine.add_state(StartupState())
